Update unit tests
