+++ src/@next/components/organisms/CheckoutAddress/fixtures.ts
@@ -61,6 +61,7 @@ const countries: GetShop_shop_countries[] = [
 ];
 
 export const LOGGED_IN_USER_PROPS = {
+  billingAsShippingPossible: true,
   countries,
   userAddresses: [
     {
@@ -70,6 +71,8 @@ export const LOGGED_IN_USER_PROPS = {
 };
 
 export const ANONYMOUS_USER_PROPS = {
-  checkoutAddress: formAddress,
+  billingAsShippingPossible: true,
+  checkoutShippingAddress: formAddress,
+  checkoutBillingAddress: formAddress,
   countries,
 };
