+++ src/@next/components/organisms/CheckoutAddress/test.tsx
@@ -11,11 +11,16 @@ import { ANONYMOUS_USER_PROPS, LOGGED_IN_USER_PROPS } from "./fixtures";
 describe("<CheckoutAddress />", () => {
   it("renders user addresses", () => {
     const setShippingAddress = jest.fn();
+    const setBillingAddress = jest.fn();
+    const setBillingAsShippingAddress = jest.fn();
     const wrapper = mount(
       <IntlProvider locale="en">
         <CheckoutAddress
           {...LOGGED_IN_USER_PROPS}
+          shippingAddressRequired
           setShippingAddress={setShippingAddress}
+          setBillingAddress={setBillingAddress}
+          setBillingAsShippingAddress={setBillingAsShippingAddress}
         />
       </IntlProvider>
     );
@@ -30,16 +35,21 @@ describe("<CheckoutAddress />", () => {
 
   it("renders address form", () => {
     const setShippingAddress = jest.fn();
+    const setBillingAddress = jest.fn();
+    const setBillingAsShippingAddress = jest.fn();
     const wrapper = mount(
       <IntlProvider locale="en">
         <CheckoutAddress
           {...ANONYMOUS_USER_PROPS}
+          shippingAddressRequired
           setShippingAddress={setShippingAddress}
+          setBillingAddress={setBillingAddress}
+          setBillingAsShippingAddress={setBillingAsShippingAddress}
         />
       </IntlProvider>
     );
 
-    const address = ANONYMOUS_USER_PROPS.checkoutAddress;
+    const address = ANONYMOUS_USER_PROPS.checkoutShippingAddress;
     const getValue = (n: number) => wrapper.find(Input).at(n).prop("value");
     expect(getValue(0)).toEqual(address.firstName);
     expect(getValue(1)).toEqual(address.lastName);
