+++ src/@next/components/organisms/CheckoutPayment/stories.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import { IntlProvider } from "react-intl";
 
 import { CheckoutPayment } from ".";
-import { ANONYMOUS_USER_PROPS, LOGGED_IN_USER_PROPS } from "./fixtures";
+import { PROPS } from "./fixtures";
 
 let portalRoot = document.getElementById("modal-root");
 if (!portalRoot) {
@@ -13,10 +13,6 @@ if (!portalRoot) {
   document.body.appendChild(portalRoot);
 }
 
-const setBillingAddress = action("setBillingAddress has been called");
-const setBillingAsShippingAddress = action(
-  "setBillingAsShippingAddress has been called"
-);
 const addPromoCode = action("addPromoCode has been called");
 const removeVoucherCode = action("removeVoucherCode has been called");
 const submitUnchangedDiscount = action(
@@ -31,24 +27,7 @@ storiesOf("@components/organisms/CheckoutPayment", module)
   .add("default", () => (
     <IntlProvider locale="en">
       <CheckoutPayment
-        {...ANONYMOUS_USER_PROPS}
-        setBillingAddress={setBillingAddress}
-        setBillingAsShippingAddress={setBillingAsShippingAddress}
-        addPromoCode={addPromoCode}
-        removeVoucherCode={removeVoucherCode}
-        submitUnchangedDiscount={submitUnchangedDiscount}
-        selectPaymentGateway={selectPaymentGateway}
-        processPayment={processPayment}
-        onGatewayError={onGatewayError}
-      />
-    </IntlProvider>
-  ))
-  .add("with addresses", () => (
-    <IntlProvider locale="en">
-      <CheckoutPayment
-        {...LOGGED_IN_USER_PROPS}
-        setBillingAddress={setBillingAddress}
-        setBillingAsShippingAddress={setBillingAsShippingAddress}
+        {...PROPS}
         addPromoCode={addPromoCode}
         removeVoucherCode={removeVoucherCode}
         submitUnchangedDiscount={submitUnchangedDiscount}
